pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "krabercode"
version = "0.1.0"
description = "A powerful CLI AI coding assistant with multi-model support and MCP integration"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "KraberCode Team" }
]
keywords = [
    "ai",
    "coding-assistant",
    "cli",
    "llm",
    "mcp",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "litellm>=1.40.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.0",
    "anyio>=3.7.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0;python_version>='3.10'",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "black>=23.0.0",
]

[project.scripts]
krabercode = "krabercode.__main__:main"
kc = "krabercode.__main__:main"

[project.urls]
Homepage = "https://github.com/krabercode/krabercode"
Documentation = "https://github.com/krabercode/krabercode#readme"
Repository = "https://github.com/krabercode/krabercode"
Issues = "https://github.com/krabercode/krabercode/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/krabercode"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.lint.isort]
known-first-party = ["krabercode"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"